BUG: make unset_value the sentinel instance, not the type
`cloudpickle_generators.unset_value` was being set to the `unset_value_type`,
not the `unset_value` sentinel instance. Because
`cloudpickle._core.unset_value_type` didn't exist (and refer to
`unset_value_type`), cloudpickle thought this was a dynamic class and serialized
the class and all of it's members. This worked until a recent change to
cloudpickle upstream, but was likely much less efficient than it should have
been.

Now, `cloudpickle._core.unset_value` properly refers to the unset value
singleton, which implements `__reduce__` as `return
'cloudpickle._core.unset_value`, so it should serialize as a simple string and
import.