Add python3.8 support
This is partly based on
https://github.com/llllllllll/cloudpickle-generators/pull/14

This fixes the test_self_in_closure test, but one async_generator test
still fails. That is because the way we used to drive the async
generator for testing no longer works in 3.8, even before the
cloudpickle round trip. I've marked it as xfail for now.